Simplify serde
